requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
audioop-lts==0.2.2
beautifulsoup4==4.13.5
blinker==1.9.0
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
diffusers==0.34.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
faiss-cpu==1.12.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
fsspec==2025.7.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
gradio==5.42.0
gradio_client==1.11.1
groovy==0.1.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.1.1
networkx==3.3
numpy==2.3.2
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-whois==0.9.5
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
ruff==0.12.9
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
sentence-transformers==5.1.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
starlette==0.47.2
streamlit==1.48.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
zipp==3.23.0
